move universal autolog to fluent module
