+++ mlflow/__init__.py
@@ -32,7 +32,6 @@ import logging
 
 from mlflow.version import VERSION as __version__  # pylint: disable=unused-import
 from mlflow.utils.logging_utils import _configure_mlflow_loggers
-from mlflow.utils.autologging_utils import universal_autolog
 import mlflow.tracking._model_registry.fluent
 import mlflow.tracking.fluent
 
@@ -109,6 +108,7 @@ set_tags = mlflow.tracking.fluent.set_tags
 delete_experiment = mlflow.tracking.fluent.delete_experiment
 delete_run = mlflow.tracking.fluent.delete_run
 register_model = mlflow.tracking._model_registry.fluent.register_model
+autolog = mlflow.tracking.fluent.universal_autolog
 
 
 run = projects.run
@@ -143,6 +143,7 @@ __all__ = [
     "get_registry_uri",
     "set_registry_uri",
     "list_run_infos",
+    "autolog",
     # model flavors
     "fastai",
     "gluon",
@@ -159,5 +160,3 @@ __all__ = [
     "tensorflow",
     "xgboost",
 ]
-
-autolog = universal_autolog
