+++ mlflow/utils/autologging_utils.py
@@ -3,8 +3,6 @@ import functools
 import gorilla
 import warnings
 
-from wrapt import register_post_import_hook
-
 import mlflow
 
 INPUT_EXAMPLE_SAMPLE_ROWS = 5
@@ -13,19 +11,6 @@ FAILED_MODEL_SIGNATURE_PREFIX_TEXT = "Failed to infer model signature: "
 ENSURE_AUTOLOGGING_ENABLED_TEXT = (
     "please ensure that autologging is enabled before constructing the dataset."
 )
-
-# Mapping of library module name to mlflow module name
-# eg: mxnet.gluon is the actual library, mlflow.gluon is our integration code for it
-AUTOLOG_INTEGRATIONS = {
-    "tensorflow": "tensorflow",
-    "keras": "keras",
-    "mxnet.gluon": "gluon",
-    "xgboost": "xgboost",
-    "lightgbm": "lightgbm",
-    "pyspark": "spark",
-    "sklearn": "sklearn",
-    "fastai": "fastai",
-}
 
 
 def try_mlflow_log(fn, *args, **kwargs):
@@ -194,26 +179,3 @@ def handle_input_example_and_signature(
         logger.warning(model_signature_user_msg)
 
     return input_example if log_input_example else None, model_signature
-
-
-def universal_autolog(
-    log_input_example=False, log_model_signature=True
-):  # pylint: disable=unused-argument
-    # getargvalues isnt actually deprecated
-    # https://docs.python.org/3/library/inspect.html#inspect.getargvalues
-    arg_info = inspect.getargvalues(inspect.currentframe())  # pylint: disable=deprecated-method
-    arg_values = {k: v for k, v in arg_info.locals.items() if k in arg_info.args}
-
-    def setup_autologging(module):
-        integration_module_obj = getattr(mlflow, AUTOLOG_INTEGRATIONS[module.__name__])
-        autolog_fn = getattr(integration_module_obj, "autolog")
-        needed_params = list(inspect.signature(autolog_fn).parameters.keys())
-        filtered = {k: v for k, v in arg_values.items() if k in needed_params}
-
-        autolog_fn(**filtered)
-
-    # for each autolog library, register a post-import hook.
-    # this way, we do not send any errors to the user until we know they are using the library.
-    # the post-import hook also retroactively activates for previously-imported libraries.
-    for module in AUTOLOG_INTEGRATIONS.keys():
-        register_post_import_hook(setup_autologging, module)
