+++ tests/utils/test_autologging_utils.py
@@ -253,42 +253,3 @@ def test_avoids_inferring_signature_if_not_needed(logger):
 
     assert x["data"] == 0
     logger.warning.assert_not_called()
-
-
-@pytest.mark.large
-def test_universal_autolog_calls_specific_autologs_correctly():
-    integrations_with_config = ["xgboost", "lightgbm", "sklearn"]
-
-    for integration_name in AUTOLOG_INTEGRATIONS.keys():
-        # modify the __signature__ of the mock to contain the needed parameters
-        args = (
-            {"log_input_example": bool, "log_model_signature": bool}
-            if integration_name in integrations_with_config
-            else {}
-        )
-        params = [
-            inspect.Parameter(param, inspect.Parameter.POSITIONAL_OR_KEYWORD, annotation=type_)
-            for param, type_ in args.items()
-        ]
-
-        # create the mock
-        mock = create_autospec(
-            getattr(getattr(mlflow, AUTOLOG_INTEGRATIONS[integration_name]), "autolog")
-        )
-        mock.__signature__ = inspect.Signature(params)
-
-        setattr(getattr(mlflow, AUTOLOG_INTEGRATIONS[integration_name]), "autolog", mock)
-        autolog_fn = getattr(getattr(mlflow, AUTOLOG_INTEGRATIONS[integration_name]), "autolog")
-        autolog_fn.assert_not_called()
-
-    universal_autolog(log_input_example=True, log_model_signature=True)
-
-    for integration_name in AUTOLOG_INTEGRATIONS.keys():
-        autolog_fn = getattr(getattr(mlflow, AUTOLOG_INTEGRATIONS[integration_name]), "autolog")
-
-        importlib.__import__(integration_name)
-
-        if integration_name in integrations_with_config:
-            autolog_fn.assert_called_once_with(log_input_example=True, log_model_signature=True)
-        else:
-            autolog_fn.assert_called_once_with()
